refactor(commands): extract lock domain into commands/lock.rs (lock-critical, verbatim)
Pure MOVE, zero behavior change. The 8 lock-domain Tauri commands (the
verify-before-destroy CALLERS) + their inners lift out of commands/mod.rs
into a new commands/lock.rs (#[path]-aliased lock_commands); mod.rs
33,652 -> 32,712 (-940). lib.rs generate_handler! BYTE-FOR-BYTE UNCHANGED.

Moved: lock_folder(+inner), unlock_folder, relock_folder, relock_all(+inner),
remove_lock(+inner), discard_unrecoverable_folder_lock(+inner), unlock_meeting,
discard_unrecoverable_meeting_lock. #[tauri::command] conserved 140 = trunk.

The verify-before-destroy ordering in each command is unchanged
character-for-character (proven by full-file reconstruction byte-identity):
- lock_folder/seal_folder_extras: seal each blob -> verify decryptable -> only
  then blank the plaintext column / delete the vault .md; audio: encrypt_file
  (verify-before-destroy inside) -> only then remove the plaintext WAV.
- unlock_folder/unlock_meeting: KEK -> unwrap CK -> decrypt transcript+timeline
  back -> materialize session WAV -> add folder to unlock set.
- relock_*: re-blank plaintext + drop the decrypted session WAV; .enc + *_blob stay.
- remove_lock: decrypt everything back to plaintext + re-export the .md.

The shared seal/unseal/audio/AAD helper web (seal_folder_extras, the aad_*
family, meeting_is_unlocked, folder_is_unlocked, masked_detail, sealed_document_blob,
lifecycle_guard, session_folder_ck, assert_in_vault, vault_path, ...) STAYS in
mod.rs (deeply shared with the retained move_note/delete_folder/delete_meeting
clusters); lock.rs reaches them via super::. 15 retained-private helpers the
moved commands call were promoted fn -> pub(crate) (bodies byte-identical).

Tests all stayed in mod.rs: #[test] 344 (= trunk); #[cfg(test)] mods 9.
Verified: cargo test --lib lock 175/0 (incl seal_transcript_timeline /
remove_lock_re_plaintexts / master_seal_stage round-trips /
topic_index_refuses_write_when_sealed_at_rest_mid_flight), commands:: 351/0
(incl masked_detail_nulls_audio_path asset-protocol gate, write_after_relock_is_refused,
verify_commands_refuse_a_sealed_meeting, assert_in_vault_rejects_symlink_escape),
clippy --lib -D warnings clean, MSRV grep clean.
